+++ docassemble_base/docassemble/base/functions.py
@@ -33,7 +33,7 @@ from user_agents import parse as ua_parse
 import phonenumbers
 locale.setlocale(locale.LC_ALL, '')
 
-__all__ = ['alpha', 'roman', 'item_label', 'ordinal', 'ordinal_number', 'comma_list', 'word', 'get_language', 'set_language', 'get_dialect', 'set_country', 'get_country', 'get_locale', 'set_locale', 'comma_and_list', 'need', 'nice_number', 'quantity_noun', 'currency_symbol', 'verb_past', 'verb_present', 'noun_plural', 'noun_singular', 'indefinite_article', 'capitalize', 'space_to_underscore', 'force_ask', 'period_list', 'name_suffix', 'currency', 'static_image', 'title_case', 'url_of', 'process_action', 'url_action', 'get_info', 'set_info', 'get_config', 'prevent_going_back', 'qr_code', 'action_menu_item', 'from_b64_json', 'defined', 'value', 'message', 'response', 'json_response', 'command', 'background_response', 'background_response_action', 'single_paragraph', 'quote_paragraphs', 'location_returned', 'location_known', 'user_lat_lon', 'interview_url', 'interview_url_action', 'interview_url_as_qr', 'interview_url_action_as_qr', 'interview_email', 'get_emails', 'action_arguments', 'action_argument', 'get_default_timezone', 'user_logged_in', 'user_privileges', 'user_has_privilege', 'user_info', 'task_performed', 'task_not_yet_performed', 'mark_task_as_performed', 'times_task_performed', 'set_task_counter', 'background_action', 'background_response', 'background_response_action', 'us', 'set_live_help_status', 'chat_partners_available', 'phone_number_in_e164', 'phone_number_is_valid', 'countries_list', 'country_name', 'write_record', 'read_records', 'delete_record', 'variables_as_json', 'all_variables', 'language_from_browser', 'device', 'plain', 'bold', 'italic', 'subdivision_type', 'indent', 'raw', 'fix_punctuation', 'set_progress', 'get_progress', 'referring_url', 'undefine', 'dispatch', 'yesno', 'noyes', 'phone_number_part', 'log', 'encode_name', 'decode_name', 'interview_list', 'interview_menu', 'server_capabilities', 'session_tags', 'get_chat_log', 'get_user_list', 'get_user_info', 'set_user_info', 'get_user_secret', 'get_session_variables', 'set_session_variables', 'go_back_in_session']
+__all__ = ['alpha', 'roman', 'item_label', 'ordinal', 'ordinal_number', 'comma_list', 'word', 'get_language', 'set_language', 'get_dialect', 'set_country', 'get_country', 'get_locale', 'set_locale', 'comma_and_list', 'need', 'nice_number', 'quantity_noun', 'currency_symbol', 'verb_past', 'verb_present', 'noun_plural', 'noun_singular', 'indefinite_article', 'capitalize', 'space_to_underscore', 'force_ask', 'period_list', 'name_suffix', 'currency', 'static_image', 'title_case', 'url_of', 'process_action', 'url_action', 'get_info', 'set_info', 'get_config', 'prevent_going_back', 'qr_code', 'action_menu_item', 'from_b64_json', 'defined', 'value', 'message', 'response', 'json_response', 'command', 'background_response', 'background_response_action', 'single_paragraph', 'quote_paragraphs', 'location_returned', 'location_known', 'user_lat_lon', 'interview_url', 'interview_url_action', 'interview_url_as_qr', 'interview_url_action_as_qr', 'interview_email', 'get_emails', 'action_arguments', 'action_argument', 'get_default_timezone', 'user_logged_in', 'user_privileges', 'user_has_privilege', 'user_info', 'task_performed', 'task_not_yet_performed', 'mark_task_as_performed', 'times_task_performed', 'set_task_counter', 'background_action', 'background_response', 'background_response_action', 'us', 'set_live_help_status', 'chat_partners_available', 'phone_number_in_e164', 'phone_number_is_valid', 'countries_list', 'country_name', 'write_record', 'read_records', 'delete_record', 'variables_as_json', 'all_variables', 'language_from_browser', 'device', 'plain', 'bold', 'italic', 'subdivision_type', 'indent', 'raw', 'fix_punctuation', 'set_progress', 'get_progress', 'referring_url', 'undefine', 'dispatch', 'yesno', 'noyes', 'phone_number_part', 'log', 'encode_name', 'decode_name', 'interview_list', 'interview_menu', 'server_capabilities', 'session_tags', 'get_chat_log', 'get_user_list', 'get_user_info', 'set_user_info', 'get_user_secret', 'get_session_variables', 'set_session_variables', 'go_back_in_session', 'manage_privileges']
 
 # debug = False
 # default_dialect = 'us'
@@ -2705,6 +2705,9 @@ def safe_json(the_object, level=0):
     if isinstance(the_object, datetime.datetime):
         serial = the_object.isoformat()
         return serial
+    if isinstance(the_object, datetime.time):
+        serial = the_object.isoformat()
+        return serial
     if isinstance(the_object, decimal.Decimal):
         return float(the_object)
     from docassemble.base.core import DAObject
@@ -2927,6 +2930,26 @@ def get_user_list(include_inactive=False):
         return server.get_user_list(include_inactive=include_inactive)
     return None
 
+def manage_privileges(*pargs):
+    """Gets or sets information about privileges on the system."""
+    if this_thread.current_info['user']['is_authenticated']:
+        arglist = list(pargs)
+        if len(arglist) == 0:
+            command = 'list'
+        else:
+            command = arglist.pop(0)
+        if command == 'list':
+            return server.get_privileges_list()
+        elif command == 'add':
+            for priv in arglist:
+                server.add_privilege(priv)
+        elif command == 'remove':
+            for priv in arglist:
+                server.remove_privilege(priv)
+        else:
+            raise Exception("manage_privileges: invalid command")
+    return None
+
 def get_user_info(user_id=None, email=None):
     """Returns information about the given user, or the current user, if no user ID or e-mail is provided."""
     if this_thread.current_info['user']['is_authenticated']:
@@ -2940,7 +2963,9 @@ def set_user_info(**kwargs):
     user_id = kwargs.get('user_id', None)
     email = kwargs.get('email', None)
     server.set_user_info(**kwargs)
-    if (user_id is None and email is None) or (user_id is not None and user_id == this_thread.current_info['user']['id']) or (email is not None and email == this_thread.current_info['user']['email']):
+    if 'privileges' in kwargs and type(kwargs['privileges']) in (list, tuple) and len(kwargs['privileges']) > 0:
+        this_thread.current_info['user']['roles'] = [y for y in kwargs['privileges']]
+    if (user_id is None and email is None) or (user_id is not None and user_id == this_thread.current_info['user']['theid']) or (email is not None and email == this_thread.current_info['user']['email']):
         for key, val in kwargs.iteritems():
             if key in ('first_name', 'last_name', 'country', 'subdivisionfirst', 'subdivisionsecond', 'subdivisionthird', 'organization', 'timezone', 'language'):
                 this_thread.current_info['user'][key] = val
@@ -2960,7 +2985,7 @@ def set_session_variables(yaml_filename, session_id, variables, secret=None):
     """Sets variables in the interview dictionary for the given interview session."""
     if session_id == get_uid() and yaml_filename == this_thread.current_info.get('yaml_filename', None):
         raise Exception("You cannot set variables in the current interview session")
-    server.set_session_variables(yaml_filename, session_id, variables, secret=secret)
+    server.set_session_variables(yaml_filename, session_id, dict(), secret=secret, literal_variables=variables)
 
 def go_back_in_session(yaml_filename, session_id, secret=None):
     """Goes back one step in an interview session."""