+++ docassemble_webapp/docassemble/webapp/cron.py
@@ -14,7 +14,7 @@ if __name__ == "__main__":
         remaining_arguments.append(arguments.pop(0))
     import docassemble.base.config
     docassemble.base.config.load(arguments=remaining_arguments)
-from docassemble.webapp.server import UserModel, UserDict, logmessage, unpack_dictionary, db, set_request_active, fetch_user_dict, save_user_dict, fresh_dictionary, reset_user_dict, obtain_lock, release_lock, app
+from docassemble.webapp.server import UserModel, UserDict, logmessage, unpack_dictionary, db, set_request_active, fetch_user_dict, save_user_dict, fresh_dictionary, reset_user_dict, obtain_lock, release_lock, app, login_user, get_user_object
 import docassemble.webapp.backend
 import docassemble.base.interview_cache
 import docassemble.base.parse
@@ -72,60 +72,63 @@ def run_cron(cron_type):
         if 'allow_cron' in the_dict:
             if the_dict['allow_cron']:
                 to_do.append(dict(key=record.key, user_id=record.user_id, filename=record.filename))
-    for item in to_do:
-        #sys.stderr.write("Doing " + str(item['key']) + " for " + str(item['filename']) + "\n")
-        try:
-            #sys.stderr.write("  " + str(cron_type) + " get interview\n")
-            interview = docassemble.base.interview_cache.get_interview(item['filename'])
-        except:
-            continue
-        for cron_type_to_use in cron_types:
-            if cron_type_to_use not in interview.questions:
-                continue
-            if re.search(r'_background$', cron_type_to_use):
-                #sys.stderr.write("  Spawning bg task: " + str(cron_type_to_use) + "\n")
-                new_task = docassemble.webapp.worker.background_action.delay(item['filename'], user_info, item['key'], None, None, None, {'action': cron_type_to_use, 'arguments': dict()})
-            else:
+    with app.app_context():
+        with app.test_request_context():
+            login_user(cron_user, remember=False)
+            for item in to_do:
+                # sys.stderr.write("Doing " + str(item['key']) + " for " + str(item['filename']) + "\n")
                 try:
-                    #sys.stderr.write("  " + str(cron_type_to_use) + " status\n")
-                    docassemble.base.functions.reset_local_variables()
-                    obtain_lock(item['key'], item['filename'])
-                    steps, user_dict, is_encrypted = fetch_user_dict(item['key'], item['filename'])
-                    interview_status = docassemble.base.parse.InterviewStatus(current_info=dict(user=dict(is_anonymous=False, is_authenticated=True, email=cron_user.email, theid=cron_user.id, the_user_id=cron_user.id, roles=[role.name for role in cron_user.roles], firstname=cron_user.first_name, lastname=cron_user.last_name, nickname=cron_user.nickname, country=cron_user.country, subdivisionfirst=cron_user.subdivisionfirst, subdivisionsecond=cron_user.subdivisionsecond, subdivisionthird=cron_user.subdivisionthird, organization=cron_user.organization, location=None), session=item['key'], yaml_filename=item['filename'], url=None, url_root=None, encrypted=is_encrypted, action=cron_type_to_use, interface='cron', arguments=dict()))
-                    #sys.stderr.write("  " + str(cron_type_to_use) + " fetch\n")
-                    #sys.stderr.write("  " + str(cron_type_to_use) + " assemble\n")
-                    interview.assemble(user_dict, interview_status)
-                    #sys.stderr.write("  " + str(cron_type_to_use) + " save\n")
-                    if interview_status.question.question_type in ["restart", "exit", "logout"]:
-                        #sys.stderr.write("  Deleting dictionary\n")
-                        reset_user_dict(item['key'], item['filename'])
-                        release_lock(item['key'], item['filename'])
-                        #sys.stderr.write("  Deleted dictionary\n")
-                    elif interview_status.question.question_type in ["backgroundresponseaction"]:
-                        #sys.stderr.write("  Got background response action\n")
-                        new_action = interview_status.question.action
-                        interview_status = docassemble.base.parse.InterviewStatus(current_info=dict(user=user_info, session=item['key'], secret=None, yaml_filename=item['filename'], url=None, url_root=None, encrypted=is_encrypted, action=new_action['action'], arguments=new_action['arguments'], interface='cron'))
-                        try:
-                            interview.assemble(user_dict, interview_status)
-                            #sys.stderr.write("Assembled 2 ok\n")
-                        except:
-                            #sys.stderr.write("Assembled 2 not ok\n")
-                            pass
-                        save_user_dict(item['key'], user_dict, item['filename'], encrypt=False, manual_user_id=cron_user.id, steps=steps)
-                        release_lock(item['key'], item['filename'])
-                    else:
-                        #sys.stderr.write("  Saving where type is " + str(cron_type_to_use) + "\n")
-                        save_user_dict(item['key'], user_dict, item['filename'], encrypt=False, manual_user_id=cron_user.id, steps=steps)
-                        release_lock(item['key'], item['filename'])
-                        if interview_status.question.question_type == "response":
-                            if hasattr(interview_status.question, 'all_variables'):
-                                sys.stdout.write(docassemble.webapp.backend.dict_as_json(user_dict).encode('utf8') + "\n")
-                            elif not hasattr(interview_status.question, 'binaryresponse'):
-                                sys.stdout.write(interview_status.questionText.rstrip().encode('utf8') + "\n")
-                except Exception as err:
-                    sys.stderr.write(str(err.__class__.__name__) + ": " + str(err))
-                    release_lock(item['key'], item['filename'])
+                    #sys.stderr.write("  " + str(cron_type) + " get interview\n")
+                    interview = docassemble.base.interview_cache.get_interview(item['filename'])
+                except:
                     continue
+                for cron_type_to_use in cron_types:
+                    if cron_type_to_use not in interview.questions:
+                        continue
+                    if re.search(r'_background$', cron_type_to_use):
+                        #sys.stderr.write("  Spawning bg task: " + str(cron_type_to_use) + "\n")
+                        new_task = docassemble.webapp.worker.background_action.delay(item['filename'], user_info, item['key'], None, None, None, {'action': cron_type_to_use, 'arguments': dict()})
+                    else:
+                        try:
+                            # sys.stderr.write("  " + str(cron_type_to_use) + " status\n")
+                            docassemble.base.functions.reset_local_variables()
+                            obtain_lock(item['key'], item['filename'])
+                            steps, user_dict, is_encrypted = fetch_user_dict(item['key'], item['filename'])
+                            interview_status = docassemble.base.parse.InterviewStatus(current_info=dict(user=dict(is_anonymous=False, is_authenticated=True, email=cron_user.email, theid=cron_user.id, the_user_id=cron_user.id, roles=[role.name for role in cron_user.roles], firstname=cron_user.first_name, lastname=cron_user.last_name, nickname=cron_user.nickname, country=cron_user.country, subdivisionfirst=cron_user.subdivisionfirst, subdivisionsecond=cron_user.subdivisionsecond, subdivisionthird=cron_user.subdivisionthird, organization=cron_user.organization, location=None), session=item['key'], yaml_filename=item['filename'], url=None, url_root=None, encrypted=is_encrypted, action=cron_type_to_use, interface='cron', arguments=dict()))
+                            # sys.stderr.write("  " + str(cron_type_to_use) + " fetch\n")
+                            # sys.stderr.write("  " + str(cron_type_to_use) + " assemble\n")
+                            interview.assemble(user_dict, interview_status)
+                            # sys.stderr.write("  " + str(cron_type_to_use) + " save\n")
+                            if interview_status.question.question_type in ["restart", "exit", "logout"]:
+                                #sys.stderr.write("  Deleting dictionary\n")
+                                reset_user_dict(item['key'], item['filename'])
+                                release_lock(item['key'], item['filename'])
+                                #sys.stderr.write("  Deleted dictionary\n")
+                            elif interview_status.question.question_type in ["backgroundresponseaction"]:
+                                #sys.stderr.write("  Got background response action\n")
+                                new_action = interview_status.question.action
+                                interview_status = docassemble.base.parse.InterviewStatus(current_info=dict(user=user_info, session=item['key'], secret=None, yaml_filename=item['filename'], url=None, url_root=None, encrypted=is_encrypted, action=new_action['action'], arguments=new_action['arguments'], interface='cron'))
+                                try:
+                                    interview.assemble(user_dict, interview_status)
+                                    #sys.stderr.write("Assembled 2 ok\n")
+                                except:
+                                    #sys.stderr.write("Assembled 2 not ok\n")
+                                    pass
+                                save_user_dict(item['key'], user_dict, item['filename'], encrypt=False, manual_user_id=cron_user.id, steps=steps)
+                                release_lock(item['key'], item['filename'])
+                            else:
+                                # sys.stderr.write("  Saving where type is " + str(cron_type_to_use) + "\n")
+                                save_user_dict(item['key'], user_dict, item['filename'], encrypt=False, manual_user_id=cron_user.id, steps=steps)
+                                release_lock(item['key'], item['filename'])
+                                if interview_status.question.question_type == "response":
+                                    if hasattr(interview_status.question, 'all_variables'):
+                                        sys.stdout.write(docassemble.webapp.backend.dict_as_json(user_dict).encode('utf8') + "\n")
+                                    elif not hasattr(interview_status.question, 'binaryresponse'):
+                                        sys.stdout.write(interview_status.questionText.rstrip().encode('utf8') + "\n")
+                        except Exception as err:
+                            sys.stderr.write(str(err.__class__.__name__) + ": " + str(err) + "\n")
+                            release_lock(item['key'], item['filename'])
+                            continue
             
 if __name__ == "__main__":
     with app.app_context():